merge console-utils inside master

+++ src/time-operations.cpp
@@ -6,10 +6,9 @@
 // #include "pch.h" // includes precompiled header with 
 #include "folf-namespace/folf-include.h"
 
-
 int main()
 {
-    double days = 9.97879;
+    long double days = 9.97879;
     double newDays = 9.97879;
     std::cout << "num = " << folf::timeCalculation::daysToHours(&days) << std::endl;
     std::cout << "NewNum = " << newDays * 24 << std::endl;
